3-14

+++ src/dao/ArticleDaoInter.java
@@ -9,9 +9,9 @@ public interface ArticleDaoInter {
 	public List<Article> getArticleById(int uid);
 	public Article getArticleNewOne(int cid);             //返回最新创建的文章
 	public Article getArticleByAid(int aid);
-	public Long createArticle(Article art);
+	public int createArticle(Article art);
 	public int deleteArticle(int aid);
-	public Long getNewAid(int uid);
+	public int getNewAid(int uid);
 	public void addViewTimes(int aid,int orignalViewTimes);
 	public Long getAppointArticleNumber(int cid);
 	public List<Article> getArticleByCid(int cid,int locpage);
