+++ src/dao/DelArticleDaoInter.java
@@ -6,7 +6,7 @@ public interface DelArticleDaoInter {
 	public int addRow(Article art);
 	public List<Article> getDelArticle(int uid);
 	public int deleteRow(int aid);
-	public Long getNewAid(int uid);
+	public int getNewAid(int uid);
 	public Article getDelArticleOne(int aid);
 	public void deleteAllRow(int uid);
 }
